# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zenku API
  version: 1.0.0
  description: |
    Schema-driven REST API for Zenku. All CRUD endpoints are auto-generated from the
    Prisma/ZenStack schema. Model names in paths use lowercase (e.g., `product`, `task`).
servers:
  - url: http://localhost:3001
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token (valid 15 minutes)
        user:
          $ref: '#/components/schemas/User'
...
paths:
...
  /api/auth/login:
    post:
      summary: Login
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful. Sets HttpOnly refresh token cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      summary: Register a new user
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  description: Min 8 chars, at least one uppercase letter and one digit
                name:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '409':
          description: Email already in use
  /api/auth/refresh:
    post:
      summary: Refresh access token
      description: Requires a valid refresh token in the HttpOnly cookie `refresh_token`.
      security: []
      tags: [Auth]
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
...